requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.6
requests==2.31.0
python-dotenv==1.0.0
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.6
gunicorn==21.2.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Caching==2.0.2